Create QA Team Lead
Hi Roberto,

because of your experience in QA with people leader skills you came to my attention.

We are looking for a lead (tech / team) for our QA team that we are rebuilding. For this I would like to exchange ideas with you. Attached you will find more information about the position and strategy. 

I would be happy to discuss this with you and I am available for a first meeting.

Greetings from Munich,
Oliver